Move QRListener interface
