+++ app/src/main/java/com/github/houseorganizer/houseorganizer/util/interfaces/QRListener.java
@@ -1,4 +1,4 @@
-package com.github.houseorganizer.houseorganizer.image;
+package com.github.houseorganizer.houseorganizer.util.interfaces;
 
 public interface QRListener {
     void QRCodeFound(String QRCode);
